+++ CHANGELOG.md
@@ -5,15 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased : 0.1-rc2]
+
+## [Unreleased : 0.1]
+
+## [2020-07-06 : 0.1-rc2]
 
 ### Modified
-Fix bug in
+
 
 ### Added 
+- Added the choice to compute metrics thresold with the min from season or from the year
+- Documentation enhancements
 
+### Fixed
+- Various improvements
 
-## [20200614 : 0.1-rc1]
+## [2020-06-14 : 0.1-rc1]
 
 ### Added